correction for caching.
Previous to this, there was no real caching occuring in Content. each template was only cached once regardless of usage. there was no cacheid associated with the template and so multiple uses cached over top of each other, essentially disabling caching. This corrects this and fixes areas that should not be cached.