allow admin to choose to display title on page or not. refs #172.
I believe this meets the original intent of this ticket. Title display is now configurable (see page settings under title text field). This also creates a class property defining whether the template includes a title display or not (default false). This allows the Column1top layout to work (since the title is below first content area). This was a bit less trivial than I'd originally hoped, but should work...

If this completes the intent of the ticket, let's close it :-)